Fix crash when sync tries to delete branches twice, and add basic progress messages to sync

+++ src/metro/syncing.cpp
@@ -102,7 +102,10 @@ namespace metro {
     // Can create and delete branches as needed.
     void change_branch_target(const Repository& repo, const string& branchName, const OID& newTarget) {
         if (newTarget.isNull) {
-            delete_branch(repo, branchName);
+            // If this was a WIP branch it might already have been deleted when the base branch was deleted.
+            if ((branch_exists(repo, branchName))) {
+                delete_branch(repo, branchName);
+            }
         } else {
             repo.create_reference("refs/heads/" + branchName, newTarget, true);
             // Update the working dir if this is the current branch.
@@ -215,6 +218,7 @@ namespace metro {
 
         Remote origin = repo.lookup_remote("origin");
         credentials->tried = false;
+        cout << "Fetching changes..." << endl;
         origin.fetch(StrArray(), fetchOpts);
 
         map<string, RefTargets> branchTargets;
@@ -258,9 +262,11 @@ namespace metro {
 
                 switch (syncType) {
                     case PUSH:
+                        cout << "Pushing " << branchName << endl;
                         pushRefspecs.push_back(make_push_refspec(branchName, targets.local.isNull));
                         break;
                     case PULL:
+                        cout << "Pulling " << branchName << endl;
                         change_branch_target(repo, branchName, targets.remote);
                         break;
                     case CONFLICT:
@@ -268,6 +274,8 @@ namespace metro {
                                 branchTargets, conflictBranchNames, pushRefspecs);
                         break;
                 }
+            } else {
+                cout << "Branch " << branchName << " is already synced" << endl;
             }
         }
 
@@ -290,6 +298,7 @@ namespace metro {
             options.callbacks.payload = &payload;
 
             credentials->tried = false;
+            cout << "Pushing changes..." << endl;
             origin.push(StrArray(pushRefspecs), options);
         }
